+++ Quotations Board Backend/DTOs/TBill/NewTbill.cs
@@ -2,5 +2,5 @@ public class NewTbill
 {
     public double Tenor { get; set; }
     public DateTime IssueDate { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
 }
